requirements.txt
Django==3.1
psycopg2-binary
Pillow
djangorestframework
python-decouple
django-phonenumber-field
phonenumbers
twilio
djangorestframework-simplejwt
drf-yasg